clients/discord/pyproject.toml
[project]
name = "multiagents-discord"
version = "0.1.0"
description = "Discord client for multiagents platform"
requires-python = ">=3.12"
dependencies = [
    "discord.py>=2.4",
    "websockets>=14.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24"]

[project.scripts]
multiagents-discord = "src.__main__:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
